Basics of multithreaded charting, gui, signals in place.
git-svn-id: svn+ssh://svn.oneunified.net/var/local/svn/private/projects/oneunified/projects/trunk/TradeFrame@967 43143c16-083f-4785-a0a0-245a1fb241a9

+++ lib/TFVuTrading/ModelChartHdf5.cpp
@@ -61,7 +61,7 @@ void ModelChartHdf5::AddChartEntries( ou::ChartDataView* pChartDataView, const o
       m_ceQuoteUpper.Append( iter->DateTime(), iter->Ask() );
       m_ceVolumeUpper.Append( iter->DateTime(), iter->AskSize() );
       m_ceQuoteLower.Append( iter->DateTime(), iter->Bid() );
-      m_ceVolumeLower.Append( iter->DateTime(), -iter->BidSize() );
+      m_ceVolumeLower.Append( iter->DateTime(), - (int) iter->BidSize() );
       m_ceQuoteSpread.Append( iter->DateTime(), iter->Ask() - iter->Bid() );
     }
     m_ceQuoteUpper.SetColour( ou::Colour::Red );
